requirements.txt
# requirements.txt

# LangChain Core Libraries
langchain==0.3.27
langchain-core==0.3.74
langchain-community==0.3.25

# LangChain Model Integrations
langchain-openai==0.3.31
langchain-ollama==0.3.7
langchain-deepseek-official==0.1.0.post1

# UI and Utilities
gradio==5.43.1
loguru==0.7.3
PyYAML==6.0.2